Fix: Crash when dragging assets between nodes area and status bar
Caused by most dragging actions which span multiple areas
(e.g.: object, collection, color, material, ...).
There is a small gap between the nodes area and the status bar that
causes the crash. Can be avoided if dragged fast enough, depending
on the polling rate of the mouse.
The direction of the drag does not affect the chance of crashing.

Within this small gap, polling functions for drag and drop operators
associated with the nodes area still get called, despite the context
being associated with the status bar.
A null guard has been added to the crashing operator,
since drop-boxes are supposed to check the required area and
region context in their poll.