Fix #137398: Transforming editbones can propagate wrong envelope radius
When tweaking envelope radii of (connected) envelope bones, we dont
always make sure that corresponding head and tail radii are in sync.
Using the `Bone Size` tool should take care of this (and it works to
some extend) but tweaking the radii through the sidepanel UI or the
Properties Editor (through RNA), then corresponding heads and tails
radii get out of sync.

Once we are in such unfortunate situation, then code in the transform
system's `recalcData_edit_armature` fails with wrong assumptions, it
propagates radii from children to parents which are unexpected / not in
use.

So one thing to do would be to add this syncing of radii to
`rna_Armature_editbone_transform_update`.
This alone would solve the "problem" in new files.

For existing files that are already out of sync we could:
- add versioning that corrects this on file load (feels rather risky at
this point, not sure about unforseen consequences)
- tweak the code in `recalcData_edit_armature` to not propagate the
"wrong" values (but scale the right ones instead)
- not do anything to old files
(this PR does the second for demonstration purposes)