Tests: Warn if running benchmark script with newer version of python
It is not obvious that the builders running the daily benchmarks use a
different version of python than what we ship with blender. Using newer
language features can prevent the builder from executing correctly.

This commit is a simple mitigation to inform developers when running the
script locally that the python version used differs from 3.9.