Fix: VSE cache invalidates all strips
Caused by checking if strip frame index falls into a range instead of
timeline frame. This was introduced in 6a39d79967 and was incorrect
solution to the issue.

Cache key had `timeline_frame` field, which was removed, because it is
not updated when strip moves. The value is now calculated from frame
index.

The issue was not very noticable especially when working with larger
strips, but the behavior was incorrect and confused me when working on
cache related features.

Pull Request: https://projects.blender.org/blender/blender/pulls/137583