Fix #137654: Cryptomatte only works with first layer
The Cryptomatte node only works with the first view layer in a setup
where a view later is called ViewLayer and another called ViewLayer.001.
This is because the code used a prefix check to match the view layer to
the Cryptomatte type name, and since ViewLayer is a prefix of the other
ViewLayer.001, the check always matches the first ViewLayer.

To fix this, we just remove that guard condition and rely on later
checks.

Pull Request: https://projects.blender.org/blender/blender/pulls/137659