Refactor: Expose `ui::nodes::wmDragNodeTreeInterface`
Replace `ui::nodes::wmDragNodeTreeInterface`
by `bke::node_interface::bNodeTreeInterfaceItemReference`, so it
is available to the Nodes Space drop-boxes.

Add reference to the node tree an interface item belongs to
in the `bNodeTreeInterfaceItemReference` struct, which enables
checking whether an item is being dragged into an editor of its
owner node tree in poll functions.