Nodes: Allow dragging Group Sockets from the panel into the nodes editor
Sockets and socket panels from the Group Sockets list in any
nodes editor's panel can now be dragged into the nodes editor,
creating a Group Input or Group Output node with only the dragged
socket or, in the case of socket panels, the panel's sockets.

When holding Shift, the added node's label is renamed to match that of
the dragged socket/panel.
When dragging a panel with a toggle socket, only its socket is dragged
by default. Holding Ctrl allows dragging the entire panel.

If a panel contains both input and output sockets, both a
Group Input and a Group Output nodes will be added.