Fix: `StrokeCache#initial_normal` does not use sampled area normal
When performing a raycast into the paint BVH inside Sculpt Mode, there
are two possible values for the cursor's normal: the current vertex's
normal, or a sampled average of the normals in the cursor radius.

The "Grab Silhouette" option of the Grab brush and the Area Normal
option of brush automasking both use an initial normal value to
determine a relative influence factor. However, this normal was only
ever using the active vertex normal, not the sampled normal, leading to
a stronger than normal falloff on areas of a mesh that have a high
variance between individual vertex normals.

To fix this, check for the existence of this sampled normal when
populating the initial value and use it if found.