Fix #135773: Grease Pencil: Applying modifier changes layer order
Rather than trying to changr the order of layers to match the evaluated
Grease Pencil geometry, keep the original layer order for now.

Generally, it's less common to change the layer order of original
layers inside a modifier. It's much more common to have the
evaluated layers and original layers in the same order, so for now,
keep the original order instead.

To correctly support this, we probably need to propagate layer groups
inside of geometry nodes. Currently, layer groups are mostly
ignored inside of evaluated data.