Python API: Move Grease Pencil stroke API to bpy_extras
This moves the Grease Pencil stroke API from `_bpy_internal`
to `bpy_extra`.
One of the main reasons for doing this is so that users can
have access to proper documentation on the blender python
API page.

Currently, only the methods are getting listed. We also need
to list class properties (getters/setters).

The links from `GreasePencilDrawing` to the class in `bpy_extras` are currently also not working.