Fix #137680: Knife constraints not accurate in orthographic view
The issue was caused by the ray origin calculated with
`ED_view3d_win_to_ray_clipped` being too far from the mesh in
orthographic view.

To resolve the inaccuracy, the ray origin is now moved closer to the
mesh bounding box.

Additionally, to further improve accuracy when snapping is not
available, a fallback point is used as the constraint point.