Geometry Nodes: Attribute Field Default Value
Extends AttributeFieldInput to hold a default value which is returned
when the field is evaluated but no matching attribute was found for the
given name. Previous behavior was to not define the value and
always return a default initialized value of [usually] 0. Limiting
ability to expose field inputs as an empty field would always
output 0 and break the node function until a valid field input
was provided by users.

Initialization of the geonode execution is updated to read the interface
inputs from the source interface tree containing the 'default' values
set in the node editor (and not the modifier panel). The default values
are then passed to the AttributeFieldInput for field socket/inputs
and used as fallback when no matching attribute is found for the
attribute name set in the modifier panel.

Changes should provide minimal functionality for controlling the default
behavior for field inputs. While it is reasonable that a design for
the entire behavior would be useful, goal for this change is to provide
the basic functionality for accessing and parsing the default values.