Fix #136508: Keep zoom levels constant when resizing VSE
We can apply the `V2D_KEEPZOOM` and `V2D_KEEPOFS_Y` to ensure this
happens. If this flag is set, we should update the view2d's `oldwinx`
and `oldwiny` just like the `V2D_KEEPASPECT` case.

Divide-by-zero is a real possibility here, so we need to avoid it by
setting `oldwinx` to `winx` in the case that it is not already
initialized, which cannot be determined from the `resize` boolean alone
(e.g. loading regions from disk on startup yields a `true` `resize`
parameter).