Fix #137877: rigid body affected by forcefield even at strength zero
Blender translates the "Dynamic" setting (`RBO_FLAG_DISABLED`) into a
bullet _kinematic state_ (see `RB_body_set_kinematic_state` /
`CF_KINEMATIC_OBJECT`). This is different from the _activation state_
(done through e.g. `RB_body_set_activation_state`) and also different
from the active/passive type. So we have (at least) three levels to
control this.

And it seems like even though a bullet body is fully intended to be
kinematic (aka animated), it still accumulates forces from the
`RB_body_apply_central_force` call. I have to be honest, I have not
tracked this down further, but it _seems_ logical to only apply forces
on a body with the "Dynamic" setting (which is the "poor-mans" solution
this PR proposes).

Tested this with activation/deactivation (with seems intact), so in my
limited tests I wasnt able to break it, but this also does not mean
much...