Sculpt: Add Manifold Boolean solver, avoid BMesh conversion for trim
Instead of joining the object's mesh and the trim mesh as a BMesh, then
running the boolean operation (which requires four conversions between
mesh and BMesh for the non-float solvers), use the same API as the mesh
boolean geometry node. The "join" case is handled separately now.

Also add the new Manifold solver to the list, and make it the default.
Though it doesn't handle non-manifold sculpt meshes, it makes sense
to make it the default because it's so much faster than the other options,
making the trim tool much more compelling.