Fix #137973: Add MaterialX version info on USD export
This change adds MaterialX version information to the exported MaterialX USD materials.

Details:

In USD 25.02, the MaterialXConfigAPI schema was introduced to allow recording the MaterialX
version used to author a material. When loading MaterialX documents into USD, this schema
is automatically applied and the associated version attribute is created on the material.

Due to how the MaterialX export works (via copying the composed MaterialX spec from a
temporary stage), the MaterialXConfigAPI needs to be explicitly applied to the final
material being exported.

This change applies this MaterialXConfigAPI schema and copies the version attribute
from the temporary MaterialX stage to the final stage.

Authored by Apple: Dan Knowlton