WIP: FFmpeg hardware encoding
A very work-in-progress code which puts together bits needed for
hardware encoding. Currently is implemented for macOS Videotoolbox
and Nvidia.

The overall idea is the following:

- Create a hardware context.
  Currently it is created with the default settings.
- Set this context in the FFmpeg using set_hwframe_ctx().
  This follows the code from the `doc/examples/vaapi_encode.c`
  from FFmpeg.
- Create software-side frame where the software puts pixels, and the
  hardware-side frame which is used by the hardware encoder.
- Transfer frame to the hardware using av_hwframe_transfer_data().
- Use the hardware frame in the encoding process.

Currently the pixel format conversion is done on the CPU, mainly
because it is not very clear how to do it on the GPU.

The patch includes changes needed for the build system and the
build system to include videotoolbox framework, and the dependency
builder to enable hardware encoding in FFmpeg.

The video toolbox SDK is available on macOS since around version 10.8
which is below our deployment target. Judging by the documentation it
seems for Nvidia encoding FFmpeg does dynamic library loading.

---

TODO:

- [ ] General cleanup, to make bits fitting better together.
- [ ] Correct assert() which are currently placed to help local development.
- [ ] Error handling.
- [ ] GPU side pixel format conversion.
- [ ] Intel and AMD support, although not required for the first version.
- [ ] Whatever is missing in the list.