Fix #131778: Grease pencil cursor radius wrong for transformed object
Cursor radius wouldnt match the drawn line if the object was not at the
origin.

When getting the radius in `grease_pencil_brush_cursor_draw`,
`DrawingPlacement::project` gives us a location in layer space.
The following `project_brush_radius` expects a worldspace location as
input though.

So to resolve, convert to worldpsace again.