WIP UI: WmWindow Monitor Index
Ghost (platform-specific) windows are created, sized, and placed using
data inside our WmWindow. This PR adds monitor_index for platforms that
require this (Wayland for sure, possibly MacOS) to support placement on
multiple monitors. This is passed along to the ghost window creation
code so that they can use this if they want. A new ghost function is
also added GHOST_GetMonitorIndex, that is used to populate this value
as the window is moved or resized. This compiles on Windows, but that
platform doesn't need to do anything with this new data.