Cycles: Move OptiX OSL Camera kernel into its own PTX module
On the one hand, this improves initialization time since we don't need to load
the full OSL module with all the shading logic if we're only using a custom
camera with SVM shading.

On the other hand, it also fixes a bug where the AO and Bevel nodes wouldn't
work when using custom cameras with SVM on OptiX.

The issue there is that those two are handled by the SHADE_SURFACE_RAYTRACE
kernel, but since that one has intersection logic, we use the OptiX-specific
kernel even if OSL shading is disabled. However, with the previous unified
OSL module, this would mean loading SHADE_SURFACE_RAYTRACE from kernel_osl.cu,
which has `#undef __SVM__` and therefore doesn't handle them correctly.

With this change, we'll use the kernels from kernel_shader_raytrace.cu in that
case, which do support SVM nodes just fine.

Since it seems that we can mix modules just fine, I'm suspecting that we could
split the modules further, instead of the current approach where modules
essentially correspond to feature set tiers and each includes the previous
one's kernels as well - but that's a separate refactor.