Cleanup: VSE: Replace remaining `sequence` references in function names
This generally does not touch any functions prefixed with `seq_` unless
it is very obviously referring to strips and not the "sequencer" as a
whole. Renames should only touch areas where it makes sense. Some
references to `seq` still exist in DNA (e.g. `act_seq` in editing,
`SequenceModifierData`, etc), which can be renamed in a future patch to
keep things smaller.

Also moves `_get` to the end of the renamed function names where
applicable for standardization.

There should be no changes to current behavior.