Fix #137909: Vulkan: Crash when allocating large vertex buffers
When allocating a large vertex buffer on NVIDIA it tried to allocate it
on the GPU and host visible. This section is limited in size (256MB).
However when allocating a large vertex buffer it should not have been
chosen.

Potential fix. It solved the issue, but still needs to verify the
consequences.