Vulkan: Remove from experimental
This changes removes Vulkan from experimental. It should be feature
parity with OpenGL with the exception of USD/Hydra.
Some backports to USD/Hydra developments are being made #133717.

This change only updates some text that mentions the state of the
backend.

Thanks for the community for testing and reporting issues!