Vulkan: Make external memory optional
This PR makes external memory optional and guards it with asserts.
There is one platform known that doesn't support external memory and
that is WOA.

External memory is currently used for GPU synchronization between Cycles
and the viewport and for OpenXR. Both usecases are not available on
these devices.

This patch just makes sure that no VMAPools are created for external
resources.