Assets: Option to disable asset libraries in the Preferences
Adds a checkbox to each asset library in the Preferences to control if it is
enabled or not. A disabled asset library will not be available in asset library
selectors, and its assets will not show up in any asset listings.

This has been a long requested feature, since not all assets and asset
libraries are relevant in every context/project, and irrelevant ones can get in
the way. It's also important for onlie asset libraries support, so users can
control which asset libraries they want to keep in sync with the online source,
see #134495.