Fix #138164: Hydra: Update MaterialX to export OpenPBR surface
Switching from Standard Surface to OpenPBR as the exported MaterialX surface.
Note, anisotropic rotation is not explicitly exposed as a parameter in OpenPBR,
instead the provided tangent vector controls the orientation, so here we use
the user supplied rotation to rotate the tangent vector around the normal.

Authored by Apple: Lee Kerley