Cycles: Adaptation to new Embree API for memory management
With the new API for memory transfers in the latest
Embree release, adaptation is necessary as using the
previous API would cause compilation errors. While the
API has changed, the logic of work remains primarily
unchanged, except for a few specific places, which may
be possible to restore to the previously used behavior.
However, this will be addressed in separate commits.