Fix #132460: Remove File Browser Path Validation
An alternative to #132500, this removes validation of saved bookmarks
and system paths in File Browser. In some circumstances the act of
checking validity of a path results in a timeout that can last many
seconds. Blender will have a very long startup time if any such items
are disconnected, invalid, offline, or removed. Even if these paths are
not used during the Blender session, like network shares used only at
work, web storage for some non-Blender use, a shortcut to a USB drive,
etc.

This PR removes all the associated code, but does keep the python
"is_valid", always return true. This does mean that offline items will
not be greyed out, and clicking on one will result in a timeout at that
time but this seems like expected behavior.