Fix #138206: Pasting nodes is ignoring offset for frame children.
The copy/paste function for nodes is offsetting the node centroid to the
mouse cursor. This was skipping direct children of selected frames, because
the node position was originally relative to the frame parent. This behavior
changed in 4bf34d95, which causes child nodes to remain stuck at their
original location.

All nodes should simply be offset now, since their positions are no longer relative.