Fix: make various mesh properties non-animatable
There are many RNA properties on the Mesh ID that were animatable, but
likely shouldn't be. For example:

- The vertex indices of edges and faces.
- The X/Y/Z mirroring toggles (effectively tool settings).

This PR marks these properties as non-animatable. The downside of doing
this is that it reduces opportunities for pranks, such as preventing
your coworkers from being able to toggle X-mirror. But that seems like
an acceptable price to pay for keeping this from happening accidentally,
and and in some cases perhaps indirectly causing more serious issues.

Note that there are many, many more erroneously animatable properties
throughout Blender. This PR narrowly addresses just those that are part
of the Mesh ID type.