Fix #138321: Movieclip strip does not blend with background
Movieclip can output buffer with smaller resolution than scene
resolution. If that happens it must be tagged as transparent.

----

Ideally we would set strip metadata, so the render transform code would do these things for us, but `StripElem`, where this data is stored is not allocated for moveclip strips. It's probably best to do this as separate refactoring.